requirements.txt
# HerbaScan Backend - Python Dependencies
# FastAPI Server for True Grad-CAM Computation

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Machine Learning
tensorflow==2.20.0
numpy>=2.0.0

# Image Processing
pillow>=11.0.0

# Utilities
python-dotenv>=1.0.0

# Optional: For better logging
loguru>=0.7.2

